fix Active Content tests
Add setup and teardown for Active Content tests to allow multiple tests to run Live content.
